+++ components/form-builder/layout/DownloadFileButton.tsx
@@ -1,28 +1,8 @@
 import React, { useCallback } from "react";
-import styled from "styled-components";
 import { useTranslation } from "next-i18next";
 
 import useTemplateStore from "../store/useTemplateStore";
-import { FancyButton } from "../panel/Button";
-
-const PrimaryButton = styled(FancyButton)`
-  padding: 10px 15px;
-  background: #26374a;
-  box-shadow: inset 0 -2px 0 #515963;
-  color: white;
-
-  &:hover:not(:disabled),
-  &:active,
-  &:focus {
-    color: #ffffff;
-    background: #1c578a;
-    box-shadow: inset 0 -2px 0 #7a8796;
-  }
-
-  &:hover:active {
-    background: #16446c;
-  }
-`;
+import { Button } from "../shared/Button";
 
 const slugify = (str: string) =>
   str
@@ -39,7 +19,7 @@ const getDate = () => {
   return date.toISOString().split("T")[0];
 };
 
-export const DownloadFileButton = () => {
+export const DownloadFileButton = ({ className }: { className?: string }) => {
   const { t } = useTranslation("form-builder");
   const { getSchema, form } = useTemplateStore();
   const downloadfile = useCallback(async () => {
@@ -59,5 +39,10 @@ export const DownloadFileButton = () => {
 
     retrieveFileBlob();
   }, [getSchema]);
-  return <PrimaryButton onClick={downloadfile}>{t("saveButton")}</PrimaryButton>;
+
+  return (
+    <Button className={className} onClick={downloadfile}>
+      {t("saveButton")}
+    </Button>
+  );
 };
