+++ components/form-builder/layout/Settings.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { useTranslation } from "next-i18next";
 import useTemplateStore from "../store/useTemplateStore";
 import useNavigationStore from "../store/useNavigationStore";
+import { Button } from "../shared/Button";
 
 const Label = ({ htmlFor, children }: { htmlFor: string; children?: JSX.Element | string }) => {
   return (
@@ -39,46 +40,6 @@ const TextInput = ({
       value={value}
       onChange={onChange}
     />
-  );
-};
-
-export const Button = ({
-  children,
-  onClick,
-  className,
-  id,
-  disabled = false,
-  "aria-label": ariaLabel = undefined,
-  theme = "primary",
-}: {
-  children?: JSX.Element | string;
-  id?: string;
-  onClick?: (e: React.MouseEvent<HTMLElement>) => void;
-  className?: string;
-  disabled?: boolean;
-  "aria-label"?: string;
-  theme?: "primary" | "secondary" | "destructive";
-}) => {
-  const themes = {
-    primary:
-      "bg-blue-dark text-white-default border-black-default hover:text-white-default hover:bg-blue-light active:text-white-default active:bg-blue-active",
-    secondary:
-      "bg-white-default text-black-default border-black-default hover:text-white-default hover:bg-gray-600 active:text-white-default active:bg-gray-500",
-    destructive:
-      "bg-red-default text-white-default border-red-default hover:bg-red-destructive hover:border-red-destructive active:bg-red-hover focus:border-blue-hover",
-  };
-
-  return (
-    <button
-      onClick={onClick}
-      className={`${className} ${themes[theme]} relative py-2 px-5 rounded-lg border-2  border-solid active:top-0.5 focus:outline-2 focus:outline-blue-focus focus:outline focus:outline-offset-2 focus:bg-blue-focus focus:text-white-default`}
-      id={id}
-      disabled={disabled}
-      aria-label={ariaLabel}
-      type="button"
-    >
-      {children}
-    </button>
   );
 };
